refactor: move main contents into app runtime
