+++ src/runtime/mod.rs
@@ -1,3 +1,4 @@
+pub mod app;
 pub mod gui;
 pub mod prelude;
 pub mod recording;
