+++ src/runtime/prelude.rs
@@ -1,3 +1,6 @@
+pub use crate::register_legacy_sketches;
+pub use crate::register_sketches;
+pub use crate::runtime::app;
 pub use crate::runtime::gui::*;
 pub use crate::runtime::recording::*;
 pub use crate::runtime::registry::*;
